Fix some signed/unsigned comparisons

+++ Source/DREAM3DLib/TestFilters/MakeDataContainer.cpp
@@ -158,7 +158,7 @@ void MakeDataContainer::dataCheck()
     VertexGeom::Pointer vertices = VertexGeom::CreateGeometry(100, "TestVertexGeom");
     SharedVertexList::Pointer test = vertices->getVertices();
     float* verts = test->getPointer(0);
-    for (size_t i=0;i<vertices->getNumberOfVertices();i++)
+    for (int64_t i=0;i<vertices->getNumberOfVertices();i++)
     {
       verts[3*i] = float(0.1 + i);
       verts[3*i+1] = float(0.2 + i);
@@ -218,7 +218,7 @@ void MakeDataContainer::execute()
 
   VertexGeom::Pointer verts = getDataContainerArray()->getDataContainer(getDataContainerName())->getGeometryAs<VertexGeom>();
 
-  for (size_t i=0;i<verts->getNumberOfVertices();i++)
+  for (int64_t i=0;i<verts->getNumberOfVertices();i++)
   {
     m_FeatureIds[i] = i;
   }
